Fix commodity rules broken by EU imports
- Removes checking children on NIG2, which is
something the EU routinely ignores.
- Simplifies checking origins to not require using
the commodity code tree.
- Removes checking description rules when the
object is deleted.
