+++ common/util.py
@@ -252,11 +252,52 @@ def resolve_path(model: Type[Model], path: str):
     return relation_path
 
 
-def get_field_tuple(
-    model: Model,
-    field_name: str,
-    default: Any = None,
-) -> Tuple[str, Any]:
+def date_ranges_overlap(a: TaricDateRange, b: TaricDateRange) -> bool:
+    """Returns true if two date ranges overlap."""
+    if a.upper and b.lower > a.upper:
+        return False
+    if b.upper and a.lower > b.upper:
+        return False
+
+    return True
+
+
+def contained_date_range(
+    date_range: TaricDateRange,
+    containing_date_range: TaricDateRange,
+    fallback: Optional[Any] = None,
+) -> Optional[TaricDateRange]:
+    """
+    Returns a trimmed contained range that is fully contained by the container
+    range.
+
+    Trimming is not eager: only the minimum amount of trimming is done to ensure
+    that the result is fully contained by the container date range.
+
+    If the two ranges do not overlap, the method returns None.
+    """
+    a = date_range
+    b = containing_date_range
+
+    if not date_ranges_overlap(a, b):
+        return fallback
+
+    start_date = None
+    end_date = None
+
+    if b.upper:
+        if a.upper is None or b.upper < a.upper:
+            end_date = b.upper
+    if b.lower > a.lower:
+        start_date = b.lower
+
+    return TaricDateRange(
+        start_date or a.lower,
+        end_date or a.upper,
+    )
+
+
+def get_field_tuple(model: Model, field_name: str) -> Tuple[str, Any]:
     """
     Get the value of the named field of the specified model.
